Make lint pass
